+++ admin/AdminUtils.ts
@@ -5,7 +5,9 @@
  * Copyright © 2025 Extremely Heavy Industries Inc.
  */
 import {XH} from '@xh/hoist/core';
-import {LocalDate} from '@xh/hoist/utils/datetime';
+import {DAYS, LocalDate} from '@xh/hoist/utils/datetime';
+import {isNumber} from 'lodash';
+import {fmtDateTimeSec} from '@xh/hoist/format';
 
 /**
  * Generate a standardized filename for an Admin module grid export, without datestamp.
@@ -20,4 +22,19 @@ export function exportFilename(moduleName: string): string {
  */
 export function exportFilenameWithDate(moduleName: string): () => string {
     return () => `${XH.appCode}-${moduleName}-${LocalDate.today()}`;
+}
+
+/**
+ * Replacer for JSON.stringify to replace timestamp properties with formatted strings.
+ */
+export function timestampReplacer(k: string, v: any) {
+    if (
+        (k.endsWith('Time') || k.endsWith('Date') || k.endsWith('Timestamp') || k == 'timestamp') &&
+        isNumber(v) &&
+        v > Date.now() - 25 * 365 * DAYS // heuristic to avoid catching smaller ms ranges
+    ) {
+        return fmtDateTimeSec(v, {fmt: 'MMM DD HH:mm:ss.SSS'});
+    }
+
+    return v;
 }