+++ utils/js/index.ts
@@ -9,6 +9,5 @@ export * from './LangUtils';
 export * from './Decorators';
 export * from './LogUtils';
 export * from './DomUtils';
-export * from './BrowserUtils';
 export * from './TestUtils';
 export * from './VersionUtils';